login authentic
done

+++ src/main/java/poly/controller/admin/DeleteProduct.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import poly.controller.SaveLogged;
 import poly.service.CategoryService;
 import poly.service.ProductService;
 
@@ -23,8 +24,20 @@ public class DeleteProduct {
 
     @GetMapping("/product/delete/{id}")
     public String delete(ModelMap model, @PathVariable(name = "id") Integer id, RedirectAttributes redirect){
-        productService.deleteById(id);
-        redirect.addFlashAttribute("success", "Deleted product successfully!");
-        return "redirect:/admin/product/list";
+        if(SaveLogged.authenticated()){
+            model.addAttribute("login",SaveLogged.USER);
+            model.addAttribute("role",SaveLogged.USER.getRole());
+            if(SaveLogged.USER.getRole() == null){
+                productService.deleteById(id);
+                redirect.addFlashAttribute("success", "Deleted product successfully!");
+                return "redirect:/admin/product/list";
+            }else{
+                model.addAttribute("message","You can not access this page");
+                return "error";
+            }
+        }else {
+            return "login";
+        }
+
     }
 }
